ipmi: add ssh option
This option ca be used to run commands through the SSH interface
(if available) of the BMC of the nodes. It relies on sshpass (and
obviously ClusterShell) to provide the password to the SSH server
of the BMC/IMM.

Using sshpass within ClusterShell requires version >= 1.7 of
ClusterShell (cf. https://github.com/cea-hpc/clustershell/issues/283
for reference). Then, explicit version in package deps has been added.

Fixes #59.