images: fix group/extra packages install order
The package group (typically core) was installed *after* the extra
packages. It is better to install it before the extra packages, as it
provides essential stuff potentially useful to other packages.

For example, this is notably the case with grub2-common and
kernel-core packages. The first provides a dropin kernel install hook
/usr/lib/kernel/install.d/20-grub.install which is used by the latter
in post (through /bin/kernel-install) to fully install the kernel in
/boot.