plugin easybuild: suppress redundancy backup
due to recursive backup, dependency software can be installed many times :-( !
So we introduce elpase time in seconds after which the backup file can be regenerated.
Default is to 300. So recent software backup (less than elapsed) time no more be
backuped!
If need, --elapse can be set to zero (0) to backup all time!